remove prod/dev differentiation in build process
