--- scripts/build.bash
@@ -1,5 +1,0 @@
-#!/bin/bash
-if [[ $PWD == */token-bridge-contracts ]];
-    then $npm_execpath run hardhat compile;
-    else $npm_execpath run hardhat:prod compile;
-fi
